requirements.txt
django>=4.0,<5.0
gunicorn>=20.0,<21.0
dj-database-url>=1.0,<2.0
whitenoise>=6.0,<7.0
psycopg2>=2.0,<3.0
djangorestframework==3.14.0